python-ext-proc-filter/requirements.txt
grpcio>=1.75.0
grpcio-tools>=1.75.0
betterproto2
